Clean up code with clippy
I noticed that I had a few clippy annotations in the code but clippy was never
being run so they weren't really being used. When I ran clippy, I saw there were
a whole bunch of issues, so this fixes all of those.

Also, I added in a rule which prevents `f32` and `f64` usage, to make sure I am
mindful in the future of using floats when I might not need to.

+++ src/dvi/interpreter.rs
@@ -179,7 +179,7 @@ where
                 ..
             } => {
                 state.add_font(
-                    *font_num as i32,
+                    *font_num,
                     &Font {
                         font_name: font_name.to_string(),
                         scale: Dimen::from_scaled_points(*scale as i32),
